cfwq

+++ rvo/Simulator.cs
@@ -327,7 +327,7 @@ namespace RVO
             {
                 Agent agent = agentList[i];
 
-                if (agent.type == (int)AgentType.GroundUnit || agent.type == AgentType.GroundJumpUnit || agent.type == AgentType.AirUnit || agent.type == AgentType.SkillUnit)
+                if (agent.type == AgentType.GroundUnit || agent.type == AgentType.GroundJumpUnit || agent.type == AgentType.AirUnit || agent.type == AgentType.SkillUnit)
                 {
                     agent.computeNeighbors();
 
@@ -339,7 +339,7 @@ namespace RVO
             {
                 Agent agent = agentList[i];
 
-                if (agent.type == (int)AgentType.GroundUnit || agent.type == AgentType.GroundJumpUnit || agent.type == AgentType.AirUnit || agent.type == AgentType.SkillUnit)
+                if (agent.type == AgentType.GroundUnit || agent.type == AgentType.GroundJumpUnit || agent.type == AgentType.AirUnit || agent.type == AgentType.SkillUnit)
                 {
                     agent.update();
                 }
@@ -354,7 +354,7 @@ namespace RVO
             {
                 Agent agent = agentList[i];
 
-                if (agent.type == (int)AgentType.GroundUnit || agent.type == AgentType.GroundJumpUnit || agent.type == AgentType.AirUnit || agent.type == AgentType.SkillUnit)
+                if (agent.type == AgentType.GroundUnit || agent.type == AgentType.GroundJumpUnit || agent.type == AgentType.AirUnit || agent.type == AgentType.SkillUnit)
                 {
                     agent.computeNeighbors();
 
@@ -368,7 +368,7 @@ namespace RVO
             {
                 Agent agent = agentList[i];
 
-                if (agent.type == (int)AgentType.GroundUnit || agent.type == AgentType.GroundJumpUnit || agent.type == AgentType.AirUnit || agent.type == AgentType.SkillUnit)
+                if (agent.type == AgentType.GroundUnit || agent.type == AgentType.GroundJumpUnit || agent.type == AgentType.AirUnit || agent.type == AgentType.SkillUnit)
                 {
                     agent.update();
